requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
Pillow>=10.0
python-decouple>=3.8
requests>=2.31
WeasyPrint>=60.0
gunicorn>=21.0
